chore: reorganize code into commands and lib
also add a command that will handle all unknown cases and suggest a command to run for the user
