+++ src/help/compare.txt
@@ -1,0 +1,2 @@
+compare [baseline] [target] [--verbose]
+  Compares two profile runs against each other
